+++ lib/src/app/controllers/navigation_controller.dart
@@ -7,20 +7,20 @@ import 'package:provider/provider.dart';
 ///
 /// Contient la route actuelle dans l'application
 class NavigationController extends ChangeNotifier {
-  String _route = "/";
+  String route = "/";
 
-  void _changerRoute(String route) {
-    _route = route;
+  void changerRoute(String route) {
+    this.route = route;
     notifyListeners();
   }
 
   /// Changer la route actuelle
-  static void changerRoute(BuildContext context, String route) {
-    Provider.of<NavigationController>(context, listen: false)._changerRoute(route);
+  static void changerView(BuildContext context, String route) {
+    Provider.of<NavigationController>(context, listen: false).changerRoute(route);
   }
 
   /// Récupérer la route actuelle
   static String getRoute(BuildContext context) {
-    return Provider.of<NavigationController>(context, listen: false)._route;
+    return Provider.of<NavigationController>(context, listen: false).route;
   }
 }