+++ lib/src/app/widgets/entete_application.dart
@@ -33,7 +33,7 @@ class _EnteteApplicationState extends State<EnteteApplication> {
           Align(
             alignment: Alignment.centerLeft,
             child: Bouton(
-              onTap: () => NavigationController.changerRoute(context, widget.routeRetour),
+              onTap: () => NavigationController.changerView(context, widget.routeRetour),
               child: Text(
                 Platform.isAndroid ? "<" : "< Retour",
                 style: TextStyle(